9.10.16 Initial Quest System
Foundation for the Quest System has been laid out, most of the work from
here on will be integrating it with the current game system. A few more
tasks have to be added and Triggers still need to be well defined. The
logic behind progressing through quests should be solid and unneeded to
be changed hopefully.