Addition of NewQuestPane and bug fixes
-Fixed issue where weapons would not display in certain chests
-Fixed issue where the XP bar in the stats screen would always show 0%
-Added NewQuestPane controller and fxml file
-Added documentation to all of the mapping classes

To Sean:
When you get a chance to look at it, all of the framework is in place
for the NewQuestPane, you just have to launch it whenever you normally
would take a quest. I added the appropriate boolean flag
"questCurrentlyDisplayed" to the GamePane as well.

Also, please code the accept and decline buttons for however the quests
should be handled if the user doesn't want it. Or should we not give
them a choice?

I'm also thinking of adding a boolean flag to each quest to say whether
or not the rewards should be revealed to the player before the quest is
over, I put a little TODO note in the controller class where we'd need
to check. Thoughts?